// server/src/models/enums.ts
export enum Role {
    OWNER = "owner",
    ADMIN = "admin",
    MEMBER = "member",
}

export enum Status {
    ACTIVE = "active",
    INACTIVE = "inactive",
}

export enum AuthType {
    PASSWORD = "password",
    OAUTH = "oauth",
    TWO_FACTOR = "2fa",
}

export enum AuthMethod {
    EMAIL = "email",
    PHONE = "phone",
    GOOGLE = "google",
    GITHUB = "github",
}